Correct config problems
Specify ethPrefix in BuildNetworkOptions.

Store subnet in JSON.

IPv4Data has incorrect values in it. Pool contains the subnet.
If a gateway is configured, Gateway does contain the gateway
CIDR. If no gateway is configured, Gateway contains the
IP range with the subnet's CIDR. This is due to a bug that
appears to be in docker/go-plugins-helpers.

Isolate IPv4Data struct usage to the driver's CreateNetwork
function.

Allow for an empty gateway.